+++ apps/argus/Cargo.toml
@@ -38,7 +38,8 @@ thiserror = "1.0.61"
 futures-locks = "0.7.1"
 async-trait = "0.1.88"
 tokio-stream = "0.1.17"
-dashmap = "5.5.3"
+dashmap = "6.1.0"
+pyth-sdk = "0.8.0"
 
 
 [dev-dependencies]
